+++ src/Emprise.Admin/Pages/Quest/Delete.cshtml.cs
@@ -2,20 +2,34 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using AutoMapper;
+using Emprise.Admin.Api;
 using Emprise.Admin.Data;
 using Emprise.Admin.Entity;
+using Emprise.Admin.Extensions;
+using Emprise.Admin.Models;
+using Emprise.Domain.Core.Enums;
+using Emprise.Domain.Core.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+using Newtonsoft.Json;
 
 namespace Emprise.Admin.Pages.Quest
 {
-    public class DeleteModel : PageModel
+    public class DeleteModel : BasePageModel
     {
-        protected readonly EmpriseDbContext _db;
-
-        public DeleteModel(EmpriseDbContext db)
+        public DeleteModel(IMudClient mudClient,
+            IMapper mapper,
+            ILogger<DeleteModel> logger,
+            EmpriseDbContext db,
+            IOptionsMonitor<AppConfig> appConfig,
+            IHttpContextAccessor httpAccessor)
+            : base(db, appConfig, httpAccessor, mapper, logger, mudClient)
         {
-            _db = db;
+
         }
 
         public QuestEntity Quest { get; set; }
@@ -37,6 +51,11 @@ namespace Emprise.Admin.Pages.Quest
             if (id > 0)
             {
                 Quest = await _db.Quests.FindAsync(id);
+                if (Quest == null)
+                {
+                    ErrorMessage = $"任务 {id} 不存在！";
+                    return Page();
+                }
                 return Page();
             }
             else
@@ -56,12 +75,28 @@ namespace Emprise.Admin.Pages.Quest
             try
             {
                 var quest = await _db.Quests.FindAsync(id);
+                if (quest == null)
+                {
+                    ErrorMessage = $"任务 {id} 不存在！";
+                    return Page();
+                }
                 _db.Quests.Remove(quest);
                 await _db.SaveChangesAsync();
+
+                await AddSuccess(new OperatorLog
+                {
+                    Type = OperatorLogType.删除任务,
+                    Content = JsonConvert.SerializeObject(quest)
+                });
             }
             catch (Exception ex)
             {
                 ErrorMessage = ex.Message;
+                await AddError(new OperatorLog
+                {
+                    Type = OperatorLogType.删除任务,
+                    Content = $"id={id}，ErrorMessage={ErrorMessage}"
+                });
                 return Page();
             }
