+++ src/Emprise.Admin/Pages/Ware/Delete.cshtml.cs
@@ -2,20 +2,32 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using AutoMapper;
+using Emprise.Admin.Api;
 using Emprise.Admin.Data;
 using Emprise.Admin.Entity;
+using Emprise.Domain.Core.Enums;
+using Emprise.Domain.Core.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+using Newtonsoft.Json;
 
 namespace Emprise.Admin.Pages.Ware
 {
-    public class DeleteModel : PageModel
+    public class DeleteModel : BasePageModel
     {
-        protected readonly EmpriseDbContext _db;
-
-        public DeleteModel(EmpriseDbContext db)
+        public DeleteModel(IMudClient mudClient,
+            IMapper mapper,
+            ILogger<DeleteModel> logger,
+            EmpriseDbContext db,
+            IOptionsMonitor<AppConfig> appConfig,
+            IHttpContextAccessor httpAccessor)
+            : base(db, appConfig, httpAccessor, mapper, logger, mudClient)
         {
-            _db = db;
+
         }
 
         public WareEntity Ware { get; set; }
@@ -56,12 +68,28 @@ namespace Emprise.Admin.Pages.Ware
             try
             {
                 var ware = await _db.Wares.FindAsync(id);
+                if (ware == null)
+                {
+                    ErrorMessage = $"物品 {id} 不存在！";
+                    return Page();
+                }
                 _db.Wares.Remove(ware);
                 await _db.SaveChangesAsync();
+
+                await AddSuccess(new OperatorLog
+                {
+                    Type = OperatorLogType.删除物品,
+                    Content = JsonConvert.SerializeObject(ware)
+                });
             }
             catch (Exception ex)
             {
                 ErrorMessage = ex.Message;
+                await AddError(new OperatorLog
+                {
+                    Type = OperatorLogType.删除物品,
+                    Content = $"id={id}，ErrorMessage={ErrorMessage}"
+                });
                 return Page();
             }
